types/general.ts -> types/style-class.ts, renamed types; moved 'isCity' to utils
