+++ src/components/atoms/index.ts
@@ -11,7 +11,7 @@ import GalleryItem from './gallery-item/gallery-item.tsx';
 import Price from './price/price.tsx';
 import OfferFeatureItem from './offer-feature-item/offer-feature-item.tsx';
 import LOGO_ICON_SIZE from './logo-icon/logo-icon-size.ts';
-import { TLogoIconSize } from './logo-icon/logo-icon-size.ts';
+import { LogoIconSizeType } from './logo-icon/logo-icon-size.ts';
 
 export {
   BookmarkIcon,
@@ -30,6 +30,6 @@ export {
 };
 
 export type {
-  TLogoIconSize,
+  LogoIconSizeType,
 };
