+++ src/components/atoms/offer-feature-item/offer-feature-item.tsx
@@ -1,4 +1,4 @@
-import {OfferFeatureType} from '../../../types';
+type OfferFeatureType = 'entire' | 'bedrooms' | 'adults';
 
 type OfferFeatureItemProps = {
   content: string;
