+++ src/components/pages/index.ts
@@ -1,7 +1,7 @@
 import FavoritesPage from './favorites/favorites.tsx';
 import LoginPage from './login/login.tsx';
 import MainPage from './main/main.tsx';
-import NotFoundPage from './not-found/not-found-page.tsx';
+import NotFoundPage from './not-found/not-found.tsx';
 import OfferPage from './offer/offer.tsx';
 
 export {
