+++ src/types/style-class.ts
@@ -1,0 +1,11 @@
+type IconsAndLabelsStyleClassType = 'place-card' | 'offer' | 'reviews';
+type OfferCardStyleClassType = 'cities' | 'near-places' | 'favorites';
+type MapStyleClassType = 'cities' | 'offer';
+type OfferFeatureClassType = 'entire' | 'bedrooms' | 'adults';
+
+export type {
+  IconsAndLabelsStyleClassType,
+  OfferCardStyleClassType,
+  OfferFeatureClassType,
+  MapStyleClassType,
+};
