+++ src/utils/index.ts
@@ -1,9 +1,11 @@
 import ScrollToTop from './scroll-to-top.ts';
 import { capitalize } from './string.ts';
 import { sortOffers } from './sort-offers.ts';
+import { isCity } from './is-city.ts';
 
 export {
   ScrollToTop,
   capitalize,
   sortOffers,
+  isCity,
 };
